feat: refactor communities module and components
- Removed the CommunitiesComponent from the main app module and created a dedicated CommunitiesModule with routing.
- Implemented lazy loading for the communities feature.
- Created CommunitiesFeaturedComponent and CommunitiesListComponent with their respective templates and styles.
- Added search functionality in CommunitiesListComponent.
- Updated the layout in CommunitiesComponent to accommodate mobile and desktop views.
- Added unit tests for the new components.
