+++ apps/commudle-admin/src/app/feature-modules/communities/communities-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { CommunitiesComponent } from './components/communities/communities.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: CommunitiesComponent,
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class CommunitiesRoutingModule {}
