+++ apps/commudle-admin/src/app/feature-modules/communities/components/communities/communities-featured/communities-featured.component.html
@@ -32,9 +32,13 @@
           <h3 class="featured-community-name">{{ featuredItems.community.name }}</h3>
           <p class="featured-community-reason">{{ featuredItems.reason }}</p>
           <div class="members-events">
-            <p class="members"><nb-icon icon="people-outline"></nb-icon>{{ featuredItems.community.members_count }}</p>
+            <p class="members">
+              <nb-icon icon="people-outline"></nb-icon>
+              {{ featuredItems.community.members_count }}
+            </p>
             <p class="events-organized">
-              <nb-icon icon="calendar-outline"></nb-icon>{{ featuredItems.community.completed_events_count }}
+              <nb-icon icon="calendar-outline"></nb-icon>
+              {{ featuredItems.community.completed_events_count }}
             </p>
           </div>
           <div class="featured-community-tags">
