+++ apps/commudle-admin/src/app/feature-modules/communities/components/communities/communities-featured/communities-featured.component.scss
