+++ apps/commudle-admin/src/app/feature-modules/communities/components/communities/communities-list/communities-list.component.scss
