+++ apps/commudle-admin/src/app/feature-modules/communities/components/communities/communities-list/communities-list.component.spec.ts
@@ -7,9 +7,8 @@ describe('CommunitiesListComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [CommunitiesListComponent]
-    })
-      .compileComponents();
+      declarations: [CommunitiesListComponent],
+    }).compileComponents();
   });
 
   beforeEach(() => {
