+++ apps/commudle-admin/src/app/feature-modules/communities/components/communities/communities.component.html
@@ -9,7 +9,7 @@
         Start your community
       </a>
     </div>
-    <app-communities-featured></app-communities-featured>
+    <commudle-communities-featured></commudle-communities-featured>
     <div class="hackathons-upcoming">
       <commudle-upcoming-hackathons></commudle-upcoming-hackathons>
     </div>
@@ -32,7 +32,7 @@
     </div>
   </div>
   <div main-column *ngIf="!isMobileView">
-    <app-communities-featured></app-communities-featured>
+    <commudle-communities-featured></commudle-communities-featured>
     <commudle-communities-list></commudle-communities-list>
   </div>
   <div right-column>
