+++ apps/commudle-admin/src/app/feature-modules/communities/components/communities/communities.component.spec.ts
@@ -1,18 +1,15 @@
-import {ComponentFixture, TestBed, waitForAsync} from '@angular/core/testing';
-import {CommunitiesComponent} from './communities.component';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommunitiesComponent } from './communities.component';
 
 describe('CommunitiesComponent', () => {
   let component: CommunitiesComponent;
   let fixture: ComponentFixture<CommunitiesComponent>;
 
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [CommunitiesComponent]
-    })
-      .compileComponents();
-  }));
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [CommunitiesComponent],
+    }).compileComponents();
 
-  beforeEach(() => {
     fixture = TestBed.createComponent(CommunitiesComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
