+++ apps/commudle-admin/src/app/feature-modules/communities/components/communities/communities.component.ts
@@ -1,0 +1,22 @@
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { FooterService } from 'apps/commudle-admin/src/app/services/footer.service';
+
+@Component({
+  selector: 'commudle-communities',
+  templateUrl: './communities.component.html',
+  styleUrls: ['./communities.component.scss'],
+})
+export class CommunitiesComponent implements OnInit, OnDestroy {
+  isMobileView: boolean;
+
+  constructor(private footerService: FooterService) {}
+
+  ngOnInit(): void {
+    this.footerService.changeFooterStatus(true);
+    this.isMobileView = window.innerWidth <= 640;
+  }
+
+  ngOnDestroy(): void {
+    this.footerService.changeFooterStatus(false);
+  }
+}
